Moved exceptions to airflow compiler
